+++ src/modules/logistics/stock/inventory/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from './useCategories';
 export * from './useInventory';
-export * from './useInventoryPage';
 export * from './useProducts';
 export * from './useStockMapper';
 export * from './useStorage';